Employees: quarantine legacy sidebar, canonical server [id] page, clean settings link
