+++ app/employees/[id]/EmployeeDetailClient.tsx
@@ -1,0 +1,2 @@
+"use client";
+export default function EmployeeDetailClient() { return null; }
