+++ app/employees/[id]/schedule/page.tsx
@@ -1,0 +1,9 @@
+import PageContainer from "@/components/PageContainer";
+export default async function EmployeeSchedulePage() {
+  return (
+    <PageContainer>
+      <h1 className="text-2xl font-bold">Employee Schedule (stub)</h1>
+      {/* TODO: Month/Week/Day calendar editor */}
+    </PageContainer>
+  );
+}
